Stable is updated

+++ Sources/Libraries/variantType/typeTraits.h
@@ -1,4 +1,4 @@
-// Copyright (C) 2014-2017 Denis Netakhin <denis.netahin@yandex.ru>, Voronetskiy Nikolay <nikolay.voronetskiy@yandex.ru>
+// Copyright (C) 2014-2017 Voronetskiy Nikolay <nikolay.voronetskiy@yandex.ru>, Denis Netakhin <denis.netahin@yandex.ru>
 //
 // This library is distributed under the MIT License. See notice at the end
 // of this file.
@@ -18,7 +18,7 @@ namespace Variant
 	{
 		typedef T Type;
 		static const char* name() { return typeid(T).name(); }
-		static bool boolean(Type val) { return val != 0; }
+		static bool boolean(Type val) { return static_cast<int>(val) != 0; }
 	};
 
 	template<> struct TypeTraits<std::string>				{ typedef std::string	Type;	static const char* name() { return "string"; }	static bool boolean(Type val) { return true; } };
@@ -41,7 +41,7 @@ namespace Variant
 
 
 
-// Copyright (C) 2014-2017 Denis Netakhin <denis.netahin@yandex.ru>, Voronetskiy Nikolay <nikolay.voronetskiy@yandex.ru>
+// Copyright (C) 2014-2017 Voronetskiy Nikolay <nikolay.voronetskiy@yandex.ru>, Denis Netakhin <denis.netahin@yandex.ru>
 // 
 // Permission is hereby granted, free of charge, to any person obtaining a copy of this software and associated 
 // documentation files (the "Software"), to deal in the Software without restriction, including without limitation
